Implemented Arduino commands and Serial protocol

+++ takahoot/board.h
@@ -5,5 +5,5 @@ struct Board {
   Trigger blue;
   Trigger yellow;
   Trigger green;
-  int tolerance;
+  byte tolerance;
 };
